project/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.3.0
openai==1.10.0
openai-whisper==20231117
spacy==3.7.2
scispacy==0.5.4
medspacy==1.0.0
transformers==4.37.0
torch==2.1.2
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1
